Replace MonadBaseControl with MonadUnliftIO.
MonadUnliftIO is a simpler abstraction that covers the part
of MonadBaseControl that is uncontroversial (i.e. not "tricky"),
namely the unwrapping of the base monad in transformer stacks
with either no or a read-only context (pretty much just IdentityT
and ReaderT) and limited to IO as the base monad.

The simpler abstraction covers the majority of use-cases, primarily
the desire for IO-centric APIs that work over any ReaderT stack on
top of IO, like exception handling, concurrency, ...

The ecosystem of libraries that previously used monad-control has
already largely migrated to unliftio. See also:

https://www.snoyman.com/blog/2018/02/conduitpocalypse