Replace control when control host is down or removed
Without doing this a graceful loss of connection with the control host is
ignored and if the whole cluster is to be restarted after that, cql-io does not
get any more events. When the last node in the cluster is restarted, cql-io
gets into an unrecoverable state where it thinks the control host is connected,
but also thinks that no hosts are reachable.