Only try to reconnect control to alive hosts
It is possible that a host marked as being down still responds to register
requests, and since it has already been marked down likely there will be no
other event to notify cql-io that the host is going down. So, it is important
that reconnections only happen with the hosts which are known to be alive.